[cmake] Move auxiliaries dmft and maxent into enclosing w2dyn module
-Move the python modules auxiliaries, dmft and maxent into an enclosing
w2dyn module
-Adjust all python imports accordingly
-Adjust CMakeList files accordingly
