+++ cthyb
@@ -10,14 +10,14 @@ import warnings
 
 import numpy as np
 
-import auxiliaries as aux
-from auxiliaries import transform as tf
-from auxiliaries import hdfout
-from auxiliaries import config
-from auxiliaries import input as aux_input
+import w2dyn.auxiliaries as aux
+from w2dyn.auxiliaries import transform as tf
+from w2dyn.auxiliaries import hdfout
+from w2dyn.auxiliaries import config
+from w2dyn.auxiliaries import input as aux_input
 
-from dmft import impurity
-from dmft import orbspin
+from w2dyn.dmft import impurity
+from w2dyn.dmft import orbspin
 
 def git_revision():
     try:
@@ -34,7 +34,7 @@ def git_revision():
 # MPI initialisation
 use_mpi = True
 if use_mpi:
-    from dmft import mpi
+    from w2dyn.dmft import mpi
     mpi_comm = mpi.MPI_COMM_WORLD
     mpi_rank = mpi_comm.Get_rank()
     mpi_size = mpi_comm.Get_size()
