+++ testsuite/maxent.tests/CMakeLists.txt
@@ -6,7 +6,7 @@ add_executable(1-getgridunit 1-getgridunit.F90)
 set (maxenttests 1-getgridunit)
 
 foreach(iter ${maxenttests})
-target_link_libraries(${iter} -Wl,${CMAKE_SOURCE_DIR}/maxent/MAXENT${F2PY_SUFFIX})
+target_link_libraries(${iter} -Wl,${CMAKE_SOURCE_DIR}/w2dyn/maxent/MAXENT${F2PY_SUFFIX})
 target_link_libraries(${iter} ${PYTHON_LIBRARIES})
 endforeach(iter)
