+++ w2dyn/auxiliaries/__init__.py
