+++ w2dyn/auxiliaries/compatibility.py
