+++ w2dyn/auxiliaries/compoundIndex.py
