+++ w2dyn/auxiliaries/configspec
