+++ w2dyn/auxiliaries/onlinestatistics.py
