+++ w2dyn/auxiliaries/quantities.py
