+++ w2dyn/dmft/_compat.py
