+++ w2dyn/dmft/gw.py
