+++ w2dyn/dmft/interaction.py
