+++ w2dyn/dmft/meta.py
