+++ w2dyn/maxent/Solver.py
